+++ kernel/src/arch/std_common/console.rs
@@ -65,7 +65,7 @@ impl awkernel_lib::console::Console for Console {
         // TODO
     }
 
-    fn irq_id(&self) -> usize {
+    fn irq_id(&self) -> u16 {
         // TODO
         0
     }
